requirements.txt
# Core API Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Web Scraping
beautifulsoup4==4.12.3
requests==2.31.0
lxml==5.1.0
html5lib==1.1

# Vector Database and Embeddings
chromadb==0.5.4
sentence-transformers==2.3.1

# LLM Integration
ollama>=0.1.6

# Utilities
loguru==0.7.2
python-dotenv==1.0.0
pyyaml==6.0.1

# Async Support
aiohttp==3.9.3
asyncio==3.4.3

# Testing
#pytest==8.0.0
pytest>=7.0.0,<8.0.0
pytest-asyncio==0.23.4
#httpx==0.26.0

# Development
black==24.1.1
flake8==7.0.0
